+++ cs/TagsCloudVisualization/Layouters/RectangleLayouters/SpiralRectangleLayouter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization.Layouters
+namespace TagsCloudVisualization.Layouters.RectangleLayouters
 {
     internal class SpiralRectangleLayouter : IRectangleLayouter
     {
